# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.broadcastRecord:
    properties:
      id:
        type: string
      sent_at:
        type: string
      status:
        type: string
      subject:
        type: string
      template_id:
        type: string
    type: object
...
  handlers.previewRequest:
    properties:
      button_text:
        type: string
      button_url:
        type: string
      content:
        type: string
      preview_text:
        type: string
      subject:
        type: string
    type: object
...
  handlers.sendRequest:
    properties:
      button_text:
        type: string
      button_url:
        type: string
      content:
        type: string
      preview_text:
        type: string
      subject:
        type: string
      template_id:
        type: string
    type: object
...
host: localhost:8088
info:
  contact: {}
  description: API de autenticacao e gerenciamento de usuarios
  title: Tron Legacy API
  version: "1.0"
paths:
...
  /admin/email-marketing/audience:
    get:
      description: Retorna estatísticas da audiência do Resend
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Resend não configurado
          schema:
            type: string
        "502":
          description: Erro ao consultar audiência
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Obter audiência de email
      tags:
      - email-marketing
  /admin/email-marketing/broadcasts:
    get:
      description: Retorna o histórico de broadcasts enviados em ordem cronológica
        reversa
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Listar broadcasts
      tags:
      - email-marketing
  /admin/email-marketing/broadcasts/{id}:
    get:
      description: Retorna detalhes de um broadcast específico
      parameters:
      - description: Broadcast ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.broadcastRecord'
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Broadcast não encontrado
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Obter broadcast por ID
      tags:
      - email-marketing
  /admin/email-marketing/send:
    post:
      consumes:
      - application/json
      description: Cria e envia um broadcast via Resend para toda a audiência
      parameters:
      - description: Dados do email
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.sendRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Dados inválidos
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Resend não configurado
          schema:
            type: string
        "502":
          description: Erro ao enviar email
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Enviar email marketing
      tags:
      - email-marketing
  /admin/email-marketing/subscribers:
    get:
      description: Retorna a lista de contatos da audiência do Resend
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Resend não configurado
          schema:
            type: string
        "502":
          description: Erro ao consultar inscritos
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Listar inscritos
      tags:
      - email-marketing
  /admin/email-marketing/subscribers/{id}:
    delete:
      description: Remove um contato da audiência do Resend
      parameters:
      - description: Contact ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: ID do contato é obrigatório
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Resend não configurado
          schema:
            type: string
        "502":
          description: Erro ao remover inscrito
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Remover inscrito
      tags:
      - email-marketing
  /admin/email-marketing/templates:
    get:
      description: Retorna os templates de email marketing disponíveis
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Listar templates de email
      tags:
      - email-marketing
  /admin/email-marketing/templates/{id}/preview:
    post:
      consumes:
      - application/json
      description: Gera um preview HTML para o template especificado
      parameters:
      - description: Template ID
        in: path
        name: id
        required: true
        type: string
      - description: Dados do preview
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.previewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Dados inválidos
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Template não encontrado
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Pré-visualizar template de email
      tags:
      - email-marketing
...
securityDefinitions:
  BearerAuth:
    description: 'Digite: Bearer {seu_token_aqui}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"